+++ js/TaskDetailsController.js
@@ -7,12 +7,14 @@ app.controller('TaskDetailsController', function ($http, $rootScope, $scope, $ui
     $scope.working = { save: false }
     
     $scope.delete = function () {
-        $http.delete($rootScope.baseUrl + "odata/Tasks(" + $scope.task.Id + ")")
-            .then(response => {
-                $scope.ok()
-            }, error => {
-                console.error(error)
-            })
+        if (confirm("Are you sure you want to delete this task?")) {
+            $http.delete($rootScope.baseUrl + "odata/Tasks(" + $scope.task.Id + ")")
+                .then(response => {
+                    $scope.ok()
+                }, error => {
+                    console.error(error)
+                })
+        }
     }
 
     $scope.save = function () {
